requirements.txt
python-dotenv~=1.0.0
python-dateutil~=2.8.2
requests~=2.31.0
pydash~=7.0.3